pointers arrays strings

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -28,7 +28,7 @@ void print_number(int n)
 
 	while(numLength)
 	{
-		_putchar(((unsigned int) n / numLength) % 10 + '0');
+		_putchar(( n / numLength) % 10 + '0');
 		numLength /= 10;
 	}
 }
